requirements.txt
flask==2.3.3
pymongo==4.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1